// src/types/formTypes.ts
export interface FormWebData {
  protocol: Partial<string> | undefined
  host: Partial<string> | undefined
  port: Partial<number> | undefined
  userName: string
  webStartMethod: Partial<number>
  configFileName: string
  webUrl: Partial<string> | undefined
  status?: Partial<string> | undefined
  pid?: Partial<number> | undefined
}